Add timeouts.py to gather task durations.
This tool queries the Swarming server for the actual run times. It prints a
dense summary of each of the steps per builder, and suggested timeout values.
The timeout values are not yet implemented but this tool will be leveraged to
enforce the timeouts.

R=phajdan.jr@chromium.org
BUG=402454

Review-Url: https://codereview.chromium.org/2486003002
Cr-Original-Commit-Position: refs/heads/master@{#431890}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b274e87c99da4150b8c4e0f2600b9b5100333142