PlzNavigate: fix UAF in PageNavigationThrottle
This CL fixes a UAF bug in PageNavigationThrottle when run with
PlzNavigate enabled. When PlzNavigate is enabled, calling
NavigationHandle::Resume or NavigationHandle::CancelDeferredNavigation
may lead to the deletion of the NavigationHandle which owns the
NavigationThrottles, so it's unsafe to access any members in
PageNavigationThrottle after calling those.

BUG=551000

Review-Url: https://codereview.chromium.org/2508913002
Cr-Original-Commit-Position: refs/heads/master@{#432512}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1648a3ff10cc7e1ab12224190fd2f9ee4e7480a7