Converts ash to use aura-mus
There is still a bunch of cleanup and removal to do, but this patch is
already so big I'm hoping to keep more cleanup to future patches. The
biggest remaining issue after this patch is z-order. By that I mean
the non-client area is at a different z-order than the frame, which
leads to some interesting artifacts. That will be addressed soon, and
I'm hoping won't block other testing.

I've landed as many other patches as I could, so this is mostly just
changes to aura, but there are some properties that needed to be
changed. I could separate that out if you prefer.

BUG=669964
TEST=covered by tests
R=jamescook@chromium.org
TBR=rockot@chromium.org (for changes to testing/buildbot)

Review-Url: https://codereview.chromium.org/2539363005
Cr-Original-Commit-Position: refs/heads/master@{#436515}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fcf18b528bc1908b9e50af6ff914b0049e693fbf