Remove about:srcdoc url conversion.
This is a reland of https://codereview.chromium.org/2494633004/

Before this CL, there was an asymetry between the renderer and the
browser. Iframe that has navigated to its srcdoc attribute got the
about:srcdoc URL in the renderer and the about::blank in the browser.
The about::srcdoc url were converted to about::blank in the RendererFrameHost.
A boolean |is_srcdoc| was created and transmitted across classes in
the browser to keep track of the "real" url, such as it could be
converted back (browser->renderer).

This led to a lot unnecessary works to keep track of this boolean and
not to forget to convert the URL back.

This CL removes the conversion and the boolean from the browser.

CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

BUG=660061
TBR=dbeam,nasko

Review-Url: https://codereview.chromium.org/2534393003
Cr-Original-Commit-Position: refs/heads/master@{#436567}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee7f43bbfe2425d9a34c4fe29bdec6609c6d8d00