DevTools: Add support for installing node and running eslint over devtools.
This CL adds a step to the 'Linux Tests' builder that will run
a step (under swarming) that runs eslint over the devtools frontend
code. eslint uses node.js, so this CL also adds a hook to gclient
to download node if CHROME_HEADLESS is present in the
environment.

R=iannucci@chromium.org, dpranke@chromium.org
BUG=642046

Review-Url: https://codereview.chromium.org/2486903002
Cr-Original-Commit-Position: refs/heads/master@{#437431}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56a90381f4d44cefeb5edb83a6227ebfa02fb885