Update logdog butler (cipd_package) to new revision.
For some bots, the git_revision of cipd_package is pinned to a
very old version (summer 2016). This causes issues when trying to
communicate with other systems expecting a newer version, and leads
to issues like:
Failed to start logdog: Missing project [LOGDOG_STREAM_PROJECT]

This CL updates the pinned revision number to a newer version,
suggested in the linked bug.

A better fix would probably be to pin this in a shared location,
making it simple to update all these at the same time, instead
of specifiying it once per target.

This CL was generated by running the following script:
git grep -l -e \
  'git_revision:3ff24775a900b675866fbcacf2a8f98a18b2a16a' \
  -- "*.json" | \
  xargs sed -ie \
    's/git_revision:3ff24775a900b675866fbcacf2a8f98a18b2a16a/git_revision:25755a2c316937ee44a6432163dc5e2f9c85cf58/'

BUG=670817

Review-Url: https://codereview.chromium.org/2574563002
Cr-Original-Commit-Position: refs/heads/master@{#438005}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ada812f0132a04edc08ff0df2e63ee1b1a4d64f6