Enable render test buildbot results on bot on chromium.android.
Attepted to enable this on an FYI bot. But that is unable to
successfully run chrome_public_test_apk due to some issue. Would
like to enable this on a bot already running chrome_public_test_apk
to see if it works as intended.

Review-Url: https://codereview.chromium.org/2626183002
Cr-Original-Commit-Position: refs/heads/master@{#443114}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 21d2ea5e341bd8f4a51e73a858a2198bcd79f0b0