xvfb.py complains if command points to a directory.
An update to xvfb.py changed its input arguments and the previously
required first one pointing to the execution directory has been removed.
This change adds makes it backwards-aware: if the first argument, now
expected to be the command, points to a directory it exists gracefully
with a helpful error message.

BUG=

Review-Url: https://codereview.chromium.org/2748123004
Cr-Original-Commit-Position: refs/heads/master@{#456957}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cac17251a3c186a7cc91e27c6de76899f162215c