Adds ash_mus_unittests to FYI bot
There are a slew of failures that need to be investigated. Rather than
early out all the tests I'm adding a filter file. This makes it easier
to retry tests as functionality is addded.

I've also cleaned up a handful of things in looking at failures.

BUG=705715
TEST=this is test change
R=msw@chromium.org

Review-Url: https://codereview.chromium.org/2803933003
Cr-Original-Commit-Position: refs/heads/master@{#462628}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ed6da57dfdcee72b4fcc5e925b6b8e6384d500f4