More Mojo FYI Browser Tests
Recently the list of mash_browser_tests was expanded. However this uncovered
several flakes on the CQ.

This change update the FYI bot to test the list. So that we can identify if
there are still flakes before relanding on the main trybots.

TBR=sky@chromium.org
TEST=mash_browser_tests
BUG=678687

Review-Url: https://codereview.chromium.org/2812243004
Cr-Original-Commit-Position: refs/heads/master@{#464512}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b62de227bd60c25f9a9b7421d628022fd2631a0c