[Bookmark suggestions] Fix the new behavior for Stable as default
After a couple of rounds of experiments with parameters for bookmark
suggestions in Stable, we selected TTL=1w and desktop_visits=true as the
best performing configuration.

This CL applies these parameters as defaults in code and specifies them
also in the testing config.

BUG=681509

Review-Url: https://codereview.chromium.org/2895933002
Cr-Original-Commit-Position: refs/heads/master@{#474234}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f11f6e2360566af1ad768a62168b8389723d325