Enable checking GN missing headers on CQ and waterfall
This relands the rest of crrev.com/2903733004.

Step "check_gn_headers" is enabled on the following bots on CQ:
- linux_chromium_dbg_ng
- linux_chromium_rel_ng

And also on the following builders on waterfall:
- Linux Builder
- Linux Builder (dbg)

Note that this script can only be added to bots that have "all"
in "additional_compile_targets".

BUG=661774, 725877

Review-Url: https://codereview.chromium.org/2910603002
Cr-Original-Commit-Position: refs/heads/master@{#477058}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00f2b7ed28e3d1d864943497d6aa865ef1cde5e2