Testing config: doc.write blocking change on slow connections only
The slow-connection version of this experiment is the only one we
are planning on shipping any time soon.

As such, we should not turn it on in finch. Turning this config on
android broke a telemetry story.

Bug: 753794
Change-Id: I18f4084a1d8335d170e6446799c3b6e8cfcf1a3e
Reviewed-on: https://chromium-review.googlesource.com/619247
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Commit-Queue: Charlie Harrison <csharrison@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#495327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 007abddf6ad7f7d3f1491217119e3e7166faaace