Enabled waterfall testing for SkipCompositingSmallScrollers
This feature is already in the code base:
https://cs.chromium.org/chromium/src/content/public/common/content_features.h?q=skipcompositing&sq=package:chromium&l=78
It was for a finch experiment conducted on Canary and Dev channel.

This CL enable this feature for perf waterfall, so that it can be ready
for experiment on Beta channel.

Bug: 684631
Change-Id: I67ca37f93b57333ef9a520d01b5e68f4e64f3d41
Reviewed-on: https://chromium-review.googlesource.com/650030
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Xida Chen <xidachen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499825}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a6c3844e4b80bc1628682de8d04a493aa57696d4