Introduce ControllableHttpResponse for tests.
It is an HTTP response that can be controlled on the current thread in
tests.  It is useful for making pauses in the middle of a response. For
instance, you can send only the headers, wait in the test for an event,
perform some actions and finally send the response's body.

Bug: 765610
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_site_isolation
Change-Id: If7c8904c3d17322d64dc1186014c29cc5860be7e
Reviewed-on: https://chromium-review.googlesource.com/666926
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504599}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 14b323009e98473053deb846c2c974e3315a45a0