Fix run multiple telemetry_benchmark script
An argument was being converted to a file object, because it was copy
pasted from another script where this was needed. We don't need to use
that argument as a file in this script though, so this CL removes the
conversion to a file object.

Change-Id: I0a750c81f86550577c62f9cfb790bbc1ab9fe048
Reviewed-on: https://chromium-review.googlesource.com/707587
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507551}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65e300ca1eb65f138c60c9c4ad7532f2e0053a5f