Create a new sandbox type for the profiling service.
The new "profiling" sandbox type is a variant of the utility sandbox type. The
only difference is that on Linux, localtime will use UTC instead of brokering a
synchronous IPC to the browser process. This is needed to prevent deadlocks
between the browser process and the profiling service, when profiling is
enabled.

For more details, see Issue 772503.

Bug: 772503, 765414
Change-Id: Id16171ea26155aed254567c61dbec05f635305a8
Reviewed-on: https://chromium-review.googlesource.com/707918
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507713}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ba08cb3faee37bbeed8249f6f29d75145cc3c34