stylus: Introduce palette welcome bubble.
Display a bubble pointing to the palette tray the first time a stylus is
ejected. Track this with a pref so the next time the stylus is ejected
the bubble does not get displayed again. The bubble gets closed if the
close button on the bubble is tapped, or if clicked somewhere outside
the bubble.

To be done in later cls:
-Support for external stylus
-Styling bubble
-Support if the palette tray is invisible and other edge cases

Test: ash_unittest PaletteTrayTestWithInternalStylus.WelcomeBubble* & PaletteWelcomeBubbleTest.*
Bug: 770960
Change-Id: Iffaeeb3f95aeda6434e95a1cc0aacfaa9e22c191
Reviewed-on: https://chromium-review.googlesource.com/700821
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Jacob Dufault <jdufault@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508729}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c4e2b1f8cf9217405c31704dea6d24a67d1ed38