Enable the layout tests on 10.13.
This CL adds support to webkitpy for 10.13-specific issues,
and adds the layout tests to the Mac10.13 bot on chromium.fyi.
For now, we'll use the same baselines for 10.13 and 10.12, which
means that we'll need to suppress the 10.13 failures.

Once we're confident that 10.13 is working, we can move it onto
the main waterfalls and start supporting baselines directly.

R=qyearsley@chromium.org
BUG=774301

Change-Id: I4607c881abb541b6bcf9a648812c487cc974f189
Reviewed-on: https://chromium-review.googlesource.com/717500
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508797}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ebaa4a19c1c5707fd8a931c2f4d359a738229eb