//tools/perf: Switch buildbot arguments to correct builder name
The one buildbot step builder used to be running the tests that a particular
bot on an android builder would run. This was confusing for several people
though, as they thought the bot itself was an android bot. This CL
changes this to create a custom bot in the sharding map, which I can edit
to control exactly what tests the builder runs.

I picked a few benchmarks which take less than 10 minutes and should get
decent coverage on different types of benchmarks, in case anything is
different between them. 

Bug: 758630
Change-Id: I2559add1fe037891630ef775fb812a4a51d6962b
Reviewed-on: https://chromium-review.googlesource.com/716871
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508839}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b902d921a5dbdfb7fde34c5ddfb442fab19f8dfa