ResourceLoadScheduler: update fieldtrial_testing_config.json
To have different study groups on beta and stable channels, update
the fieldtrial_testing_config.json to have both groups.

Only the first group is verified in the perf tests. So, I put
the group for the beta channel that fits better to HEAD code, and
is more likely to be the final parameters for the launch.

In this path, I also remove Windows specific entry because it was
needed for m61 and m62, but should not be needed for later versions.

Bug: 729954
Change-Id: I5b42174dc44d078a947a0584127facf7ba29cbfc
Reviewed-on: https://chromium-review.googlesource.com/718158
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#509009}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36160f0427ad5657e299fa932c920d1f490bc1e1