Run the layout tests on the main Win bots.
This adds the layout tests to
- Win7 Tests (1)
- Win7 Tests (dbg)(1)

which will enable them on the corresponding trybots as well
(notably win7_chromium_rel_ng), for all appropriate changes.

This does not enable them on the x64 Win bots, as we don't currently
support that configuration.

This is a reland of r508861, except that we're not adding the
tests to `Win7 Tests x64 (1)` (which didn't actually work, as noted
above).

TBR=tansell@chromium.org, qyearsley@chromium.org
BUG=524758

Change-Id: Iae049d1967a831ab3f5b4be68a11c2b16328359e
Reviewed-on: https://chromium-review.googlesource.com/750240
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Tim 'mithro' Ansell <tansell@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#513388}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32af0b0b4de23ef5bfcd7a6f067877cbff04d94c