Mojo FYI Bot: Content_Browsertest with --enable-viz
This will enable the Mojo FYI bot to run content_browsertest --enable-viz
The filter file is a list of tests which fail, as they traditionally rely on information that is now in a different process.
We'll bring more online as we create the appropriate test apis.

TBR=sky@chromium.org

TEST: content_browsertest
Change-Id: I94ecb9bef6ad7b2624ecb8ddcd5ef7d3c1875a3b
Reviewed-on: https://chromium-review.googlesource.com/748776
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#513557}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a12819cc41894a3510623065caf875b34109da30