Change test filter argument to --isolated-script-test-filter.
When retrying tryjobs without the patch, it's only feasible to pass
the test list on the command line, because on Swarming, it's difficult
(and not currently possible in recipes) to pass a temporary file when
triggering the job.

Use a double-colon ("::") separator in order to support perf
benchmarks, whose names include URLs, and necessarily single colons.

Tested with the scenarios described in the bug report.

BUG=533481

Change-Id: Ic76bc9d5f4e0ad38f67fb6ccf3f5033648a86899
Reviewed-on: https://chromium-review.googlesource.com/778124
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#518151}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a649a46124875d34a36e5537fed37a36e2d3c63a