Experimental SwReporter is no longer experimental.
Move the behaviour controlled by ExperimentalSwReporterEngine and
ExperimentalSwReporterEngineOnAllArchitectures outside the feature
checks so it is always on.

The exception is the "tag" parameter which is used in testing so must
still be set through a feature until the test scripts are updated.

Bug: 786964
Change-Id: I06bd7a599b3ab61121daca939d3df60967b3c0e6
Reviewed-on: https://chromium-review.googlesource.com/779302
Commit-Queue: Joe Mason <joenotcharles@chromium.org>
Reviewed-by: Gayane Petrosyan <gayane@chromium.org>
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#520595}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5d34ad9dc7e098dde5d66c919f08bc0f534f7c24