AppWindowApiTest.Properties: Elim flakiness
This test was assuming that calling minimize() and maximize() followed
by restore() without waiting for onMinimized or onMaximized would
trigger onRestored. This is not the case on Mash, and is also a likely
cause of flake on other platforms.

Bug: 232330,678687,794771
Change-Id: I2a6a0f1e0b527583add555ed5ab8bfdc3a073df6
Reviewed-on: https://chromium-review.googlesource.com/821310
Reviewed-by: Ben Wells <benwells@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#524110}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eefdc769289d2fe0de931f6626c94aa34d0292bd