Add Viz Browser Tests to the Mojo Windows FYI bot
We want to begin getting coverage of browser_tests with --enable_viz.
To start we would like to run them on a Windows FYI bot to find flakes.

I'll be triaging the failures found so far, and updating the filter
later on with the associated bugs.

TEST: viz_browser_tests
BUG: 794961
Change-Id: I554af1c3d308da1b7fd0496c08efa24e17bd6799
Reviewed-on: https://chromium-review.googlesource.com/822295
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#524130}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e360d9b0a4640a7392335282398dc58346a98e3