Remove 'All' meta-target and clarify comments around 'gn_all'.
'All' was the name used to build everything under GYP, and used to
keep bot recipes working for both GYP & GN. Now that GYP is gone and we
build 'all' on the majority of bots, we no longer need 'All'.

The description for 'gn_all' is also clarified, to express that it is
an internal meta-target used to reference sub-BUILD.gn files, and to
thereby make their contents part of the automatic 'all' set.

(This is a reland of e978631af95caed77815cdafcbcf6c70bc773999)

Change-Id: I23664cd79d92164b05ee8b004876b12af745d910
Reviewed-on: https://chromium-review.googlesource.com/837516
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#525887}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09fb3a937ac30ed71814ca8ab790e9218cc427f1