Coalesce events properly in RenderWidgetTargeter.
We were not colalescing events properly in RenderWidgetTargeter::
MergeEventIfPossible. As a result, when we have incorrecly replaced a
mouse-wheel event of phaseEnd in the queue, we get crashes complaining
that there's already scrolling in progress when we process a new
phaseBegan event.

Bug: 802063
Test: SitePerProcessBrowserTest.InputEventRouterWheelCoalesceTest
Change-Id: Idb42b310a7542697df601e5646fcc11bd16d9701
Reviewed-on: https://chromium-review.googlesource.com/871514
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Commit-Queue: Ria Jiang <riajiang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#529993}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6eeea9f06126e377e4e30db7e66bf323923d9e5a