[LPM] Speculative fix for Mac builds of skia proto fuzzer.
On Mac builds, it looks like calling WriteNum on a size_t was causing
the compiler to select the templated version of WriteNum rather than the
uint64_t version. Hopefully it won't do the opposite (and break on calls
to WriteNum(uint64_t) because of this fix.

Bug: 808155, 769578
Change-Id: I8babcf6da5090b9bffdbcc98b573c11c02f2e489
Reviewed-on: https://chromium-review.googlesource.com/899906
Commit-Queue: Jonathan Metzman <metzman@chromium.org>
Reviewed-by: Martin Barbella <mbarbella@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#534143}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2267d0d1c488cc65e6ea5d5221559044910301e