Reland of NetworkService: Implement URLLoader::OnAuthRequired
The initial upload patch of this CL is the same as:
https://chromium-review.googlesource.com/c/chromium/src/+/877512

This reland CL addresses comments at:
https://chromium-review.googlesource.com/c/chromium/src/+/877512/19

The initial upload patch of this CL uses empty credential to indicate
that URLRequest::CancelAuth() needs to be called, which is incorrect
because empty credentials are valid. The latest patch of this CL
contains the fix for this issue.

TBR=jam@chromium.org, tsepez@chromium.org

Bug: 783990
Cq-Include-Trybots: master.tryserver.chromium.linux:linux_mojo
Change-Id: Ifdbab56badaf3e13b5f15f5aa0d31cfb764325c6
Reviewed-on: https://chromium-review.googlesource.com/938960
Commit-Queue: Jun Cai <juncai@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: Jun Cai <juncai@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#539903}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc561f417634ce7326816b5167476003386e947b