Docked Magnifier: Fix High Contrast mode
Before this CL, high contrast was applied to the contents of
the viewport twice (once on the copied contents, and another
since the viewport is a child of the root layer), resulting
in undoing its effect.
This CL updates the "inverted" status of the viewport layer.

BUG=817431
TEST=Added test coverage.

Change-Id: I4d946e5ae22f1232db74fecf7b32311073124055
Reviewed-on: https://chromium-review.googlesource.com/942602
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#540250}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 979ef78e09f6d2ea0b253a011c76289b8a2681bb