ResourceLoadScheduler: remove stale configurations
Remove the entry "Enabled_bg_limit_8", that was for the first field trial.
Also, remove "ReportRendererPeakMemoryStats" feature, that is not
planned to be used in a launching configuration, but it was just for
gathering memory metrics until Beta.

Bug: 775849
Change-Id: Ib496db8d69b6891915cfab453ad4abc13d66e522
Reviewed-on: https://chromium-review.googlesource.com/956945
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#542275}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a3d458c7e9c8ebf21833bbaa48ee4e4159a516dc